requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.

retrying>=1.2.2,!=1.3.0 # Apache-2.0
alembic>=0.6.4
anyjson>=0.3.3
argparse
cassandra-driver>=2.1.1
croniter>=0.3.4 # MIT License
eventlet>=0.15.2
happybase>=0.5,!=0.7
iso8601>=0.1.9
jsonpath-rw>=1.2.0,<2.0
jsonschema>=2.0.0,<3.0.0
keystonemiddleware>=1.0.0
lockfile>=0.8
lxml>=2.3
msgpack-python>=0.4.0
netaddr>=0.7.12
oslo.db>=1.0.0  # Apache-2.0
oslo.config>=1.4.0  # Apache-2.0
oslo.rootwrap>=1.3.0
oslo.vmware>=0.6.0                      # Apache-2.0
PasteDeploy>=1.5.0
pbr>=0.6,!=0.7,<1.0
pecan>=0.5.0
posix_ipc
oslo.messaging>=1.4.0
oslo.middleware>=0.1.0                  # Apache-2.0
oslo.serialization>=1.0.0               # Apache-2.0
oslo.utils>=1.0.0                       # Apache-2.0
pysnmp>=4.2.1,<5.0.0
python-ceilometerclient>=1.0.6
python-glanceclient>=0.14.0
python-keystoneclient>=0.11.1
python-neutronclient>=2.3.6,<3
python-novaclient>=2.18.0
python-swiftclient>=2.2.0
pytz
PyYAML>=3.1.0
requests>=2.2.0,!=2.4.0
six>=1.7.0
SQLAlchemy>=0.8.4,<=0.8.99,>=0.9.7,<=0.9.99
sqlalchemy-migrate>=0.9.1,!=0.9.2
stevedore>=1.1.0  # Apache-2.0
tooz>=0.3 # Apache-2.0
WebOb>=1.2.3
WSME>=0.6
